+++ src/main/java/de/muenchen/mpdz/zammad/ldap/service/ZammadSyncServiceSubtreeUtil.java
@@ -2,7 +2,7 @@ package de.muenchen.mpdz.zammad.ldap.service;
 
 import de.muenchen.mpdz.zammad.ldap.domain.ZammadGroupDTO;
 import de.muenchen.mpdz.zammad.ldap.domain.ZammadUserDTO;
-import de.muenchen.mpdz.zammad.ldap.ezldapextension.EnhancedLdapUserDto;
+import de.muenchen.oss.ezldap.core.EnhancedLdapUserDto;
 import de.muenchen.mpdz.zammad.ldap.tree.LdapOuNode;
 import de.muenchen.oss.ezldap.core.LdapOuSearchResultDTO;
 import de.muenchen.oss.ezldap.core.LdapUserDTO;
